+++ hippynn/layers/indexers.py
@@ -271,3 +271,13 @@ class FuzzyHistogram(torch.nn.Module):
         x = values - self.bins
         histo = torch.exp(-((x / self.sigma) ** 2) / 4)
         return torch.flatten(histo, end_dim=1)
+    
+class SpeciesIndexer(torch.nn.Module):
+    def forward(self, values, onehot_encoding):
+        n_species = onehot_encoding.shape[1]
+        values_by_species = []
+        for i in range(n_species):
+            species_mask = onehot_encoding[:,i]
+            species_values = values[species_mask]
+            values_by_species.append(species_values)
+        return values_by_species